RaceOrganizerServer/requirements.txt
locust==0.0
Django==2.2.1
pytz==2019.1
mysqlclient==1.4.2.post1
setuptools==41.0.1
django-silk==3.0.2
locustio==0.11.0